+++ config/app.php
@@ -15,7 +15,7 @@ return [
 
     'name' => env('APP_NAME', 'Laravel'),
 
-    'version' => env('APP_VERSION', '0.4.2'),
+    'version' => env('APP_VERSION', '0.4.3'),
 
     /*
     |--------------------------------------------------------------------------